Differential Expression Analysis

The functions are designed to run DEA.

deg_caller is used to run two group differential expression analysis.

deg_batch_caller is used to run batch DEG analysis.

deg_caller(data, group, ref_group)

deg_batch_caller(data, groups, ref_group)

Arguments

data
An expression dataset in data.frame format whose rows indicate genes and columns indicate samples.

group
A character vector specifying 1 of 2 groups which samples belong to.

ref_group
A string specifying the reference group.

groups
A character vector specifying the group which samples belong to. Should at least 2 groups.

Value

A data.frame or a list of data.frame.

Examples

data <- load_example_dataset()

# Get a subset of expression dataset
ix <- c(1, 2, 61, 62)
expr <- data$expr[, ix]
group <- data$pdata$perturbagen[ix]

# Run DEG analysis
report <- deg_caller(expr, group = group, ref_group = group[3])
head(report)

# Run batch DEG analysis
reports <- deg_batch_caller(expr, groups = group, ref_group = group[3])
head(reports)

identical(report, reports[[1]])