Run GSEA

Run GSEA

runGSEA(
  data = NULL,
  res.path = getwd(),
  dirct = "up",
  n.path = 6,
  msigdb.path = NULL,
  nPerm = 1000,
  minGSSize = 10,
  maxGSSize = 500,
  p.cutoff = 0.05,
  p.adj.cutoff = NULL
)

Arguments

data
a deg data.frame

res.path
A string value to indicate the path for saving the results for functional pathways.

dirct
A string value to indicate the direction of identifying significant pathway. Allowed values contain c('up', 'down'); up means up-regulated pathway, and down means down-regulated pathway; "up" by default.

n.path
A integer value to indicate how many top pathways sorted by NES should be identified; 6 by default.

msigdb.path
A string value to indicate ABSOULUTE PATH/NAME of MSigDB file (GMT file with gene symbols) downloaded from https://www.gsea-msigdb.org/gsea/msigdb/collections.jsp#H.

nPerm
A integer value to indicate the number of permutations; 1000 by default and 10000 will be better for reproducibility.

minGSSize
A integer value to indicate minimal size of each geneSet for analyzing; 10 by default

maxGSSize
A integer value to indicate maximal size of each geneSet for analyzing; 500 by default.

p.cutoff
A numeric value to indicate the nominal p value for identifying significant pathways; pvalue < 0.05 by default.

p.adj.cutoff
A numeric value to indicate the adjusted p value for identifying significant pathways; padj < 0.05 by default.

Value

A plot.

Examples

one_report <- tcm.LoadAnalyzedDEG(2)
msigdb.path <- system.file("extdata", "c2.cp.kegg.v6.2.symbols.gmt",
  package = "TCMR", mustWork = TRUE
)
runGSEA(
  data = one_report,
  dirct = "up",
  msigdb.path = msigdb.path,
  n.path = 3,
  p.cutoff = 0.05
)