Differential Expression Analysis

The functions are designed to run DEA.

deg_caller is used to run two group differential expression analysis.

deg_batch_caller is used to run batch DEG analysis.

deg_caller(data, group, ref_group)

deg_batch_caller(data, groups, ref_group)

Arguments

data
An expression dataset in data.frame format whose rows indicate genes and columns indicate samples.

group
A character vector specifying 1 of 2 groups which samples belong to.

ref_group
A string specifying the reference group.

groups
A character vector specifying the group which samples belong to. Should at least 2 groups.

Value

A data.frame or a list of data.frame.

Examples

data <- load_example_dataset()

#> Loading dataset GSE85871...

#> Done.

# Get a subset of expression dataset
ix <- c(1, 2, 61, 62)
expr <- data$expr[, ix]
group <- data$pdata$perturbagen[ix]

# Run DEG analysis
report <- deg_caller(expr, group = group, ref_group = group[3])

#> Info: Glycyrrhizic acid vs DMSO (reference group)

#> N: DMSO:#2  Glycyrrhizic acid:#2

#> Constructing design matrix...

#> Running DEG analysis with limma...

#> Warning: Zero sample variances detected, have been offset away from zero

#> Reporting results...

#> Done.

head(report)

#>    identifier     logFC  AveExpr         t      P.Value adj.P.Val         B
#> 1:      CDH12  3.411526 1.981608  30.80083 5.441761e-05 0.2925089 1.1921795
#> 2:      NLGN1 -3.204683 1.645785 -26.29158 8.915954e-05 0.2925089 1.0903322
#> 3:       HCRT  3.181742 3.144149  22.99756 1.353037e-04 0.2925089 0.9811453
#> 4:    SOSTDC1  3.582065 1.791032  22.73374 1.402533e-04 0.2925089 0.9706183
#> 5:   SERPINB2 -2.191310 2.182794 -20.95775 1.806605e-04 0.2925089 0.8909550
#> 6:   ATP6V1G2  2.698387 3.463790  20.22094 2.019340e-04 0.2925089 0.8527498

# Run batch DEG analysis
reports <- deg_batch_caller(expr, groups = group, ref_group = group[3])

#> Info: Glycyrrhizic acid vs DMSO (reference group)

#> N: DMSO:#2  Glycyrrhizic acid:#2

#> Constructing design matrix...

#> Running DEG analysis with limma...

#> Warning: Zero sample variances detected, have been offset away from zero

#> Reporting results...

#> Done.

head(reports)

#> $`Glycyrrhizic acid:DMSO`
#>        identifier         logFC  AveExpr             t      P.Value adj.P.Val
#>     1:      CDH12  3.411526e+00 1.981608  3.080083e+01 5.441761e-05 0.2925089
#>     2:      NLGN1 -3.204683e+00 1.645785 -2.629158e+01 8.915954e-05 0.2925089
#>     3:       HCRT  3.181742e+00 3.144149  2.299756e+01 1.353037e-04 0.2925089
#>     4:    SOSTDC1  3.582065e+00 1.791032  2.273374e+01 1.402533e-04 0.2925089
#>     5:   SERPINB2 -2.191310e+00 2.182794 -2.095775e+01 1.806605e-04 0.2925089
#>    ---                                                                       
#> 12544:    ZNF280B -3.608800e-04 2.389466 -2.983343e-04 9.997800e-01 1.0000000
#> 12545:      PSMF1 -2.715000e-05 9.237362 -1.651653e-04 9.998782e-01 1.0000000
#> 12546:      CFLAR -1.104091e-05 6.116883 -9.067181e-05 9.999331e-01 1.0000000
#> 12547:   CCDC102B  0.000000e+00 0.000000  0.000000e+00 1.000000e+00 1.0000000
#> 12548:      RPE65  0.000000e+00 0.000000  0.000000e+00 1.000000e+00 1.0000000
#>                 B
#>     1:  1.1921795
#>     2:  1.0903322
#>     3:  0.9811453
#>     4:  0.9706183
#>     5:  0.8909550
#>    ---           
#> 12544: -6.5415101
#> 12545: -6.5415101
#> 12546: -6.5415101
#> 12547: -6.5415101
#> 12548: -6.5415101
#> 

identical(report, reports[[1]])

#> [1] TRUE